Move utility related code to own subdir
