+++ src/util/hash.c
@@ -21,28 +21,33 @@
  ******************************************************************************/
 
 
-#ifndef __ARGENT_INCLUDE_PLUGIN_H__
-#define __ARGENT_INCLUDE_PLUGIN_H__
+#include "../../include/argent.h"
 
-#ifdef __cplusplus
-extern "C" {
-#endif
-
-#include "./object.h"
+#include <stdint.h>
 
 
-AG_OBJECT_DECLARE(ag_plugin, AG_TYPEID_PLUGIN);
-
-extern ag_plugin        *ag_plugin_new(const char *, const char *);
-extern ag_plugin        *ag_plugin_new_local(const char *);
-
-extern ag_string        *ag_plugin_dso(const ag_plugin *);
-extern ag_string        *ag_plugin_sym(const ag_plugin *);
-extern void             *ag_plugin_hnd(const ag_plugin *);
-
-#ifdef __cplusplus
+extern ag_hash
+ag_hash_new(size_t key)
+{
+        key = (key ^ (key >> 30)) * UINT64_C(0xbf58476d1ce4e5b9);
+        key = (key ^ (key >> 27)) * UINT64_C(0x94d049bb133111eb);
+        key = key ^ (key >> 31);
+    
+        return key;
 }
-#endif
 
-#endif /* !__ARGENT_INCLUDE_PLUGIN_H__ */
+
+extern ag_hash
+ag_hash_new_str(const char *key)
+{
+        AG_ASSERT_PTR (key);
+
+        register ag_hash hash = 5381;
+        register int c;
+
+        while ((c = *key++))
+                hash = ((hash << 5) + hash) + c;
+
+        return hash;
+}
