+++ src/util/hash.h
