+++ src/util/manager.h
