+++ src/util/memblock.c
@@ -21,7 +21,7 @@
  ******************************************************************************/
 
 
-#include "../include/argent.h"
+#include "../../include/argent.h"
 
 #ifdef __FreeBSD__
 #       include <malloc_np.h>
